// frontend/xacria-presenze/src/generated-client/model/fullUserProfileResponseDto.ts
/**
 * API Presenze-Pro
 * API used by the Presenze-Pro thesis project’s client
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface FullUserProfileResponseDto { 
    name?: string;
    surname?: string;
    serialNum?: number;
    duty?: string;
    employmentType?: string;
    hireDate?: Date;
    birthDate?: Date;
    address?: string;
    phone?: string;
    iban?: number;
    email?: string;
    role?: FullUserProfileResponseDto.RoleEnum;
}
export namespace FullUserProfileResponseDto {
    export type RoleEnum = 'USER' | 'ADMIN' | 'OWNER';
    export const RoleEnum = {
        USER: 'USER' as RoleEnum,
        ADMIN: 'ADMIN' as RoleEnum,
        OWNER: 'OWNER' as RoleEnum
    };
}